requirements.txt
flask
python-docx
nltk
scikit-learn
